fix(renderer): use float arithmetic in trig/projection
Camera::Matrix was computing (m_Width / m_Height) as integer division
before the float cast, producing 1.0 for any 16:9 window — a broken
projection matrix. m_LastX/Y initializers also truncated to int.
Renderer2D circle subdivision angles had the same truncation.

+++ Onyx/Source/Graphics/Camera.cpp
@@ -8,8 +8,8 @@ namespace Onyx {
 	Camera::Camera(int width, int height, glm::vec3 position)
 		: m_Width(width), m_Height(height), m_Position(position)
 	{
-		m_LastX = width / 2;
-		m_LastY = height / 2;
+		m_LastX = width / 2.0f;
+		m_LastY = height / 2.0f;
 	}
 
 	void Camera::Matrix(float FOVdeg, float nearPlane, float farPlane, Shader& shader, const char* uniform)
@@ -20,7 +20,7 @@ namespace Onyx {
 		glm::mat4 projection = glm::mat4(1.0f);
 
 		view = GetViewMatrix();
-		projection = glm::perspective(glm::radians(FOVdeg), (float)(m_Width / m_Height), nearPlane, farPlane);
+		projection = glm::perspective(glm::radians(FOVdeg), (float)m_Width / (float)m_Height, nearPlane, farPlane);
 
 		glUniformMatrix4fv(glGetUniformLocation(shader.GetProgramID(), uniform), 1, GL_FALSE, glm::value_ptr(projection * view));
 	}
